pyproject.toml
[tool.poetry]
name = "bsuir-backend"
version = "0.1.0"
description = ""
authors = ["V-outsider <voutsider@tutanota.com>"]
readme = "README.md"
packages = [{include = "bsuir_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
py2neo = "^2021.2.4"
uvicorn = "^0.24.0.post1"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.3.1"
pandas = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
